+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.14.9] - (xx.06.2024)
+
+## Changes
+
+- refactor check JWT token
+- refactor check arf profile
+- correct some logs
+- cleanup
+
 ## [7.14.8] - (01.06.2024)
 
 ## Changes